+++ packages/node/src/ethereum/api.ethereum.test.ts
@@ -17,7 +17,8 @@ import {
 } from './block.ethereum';
 
 // Add api key to work
-const HTTP_ENDPOINT = 'https://eth.api.onfinality.io/public';
+// const HTTP_ENDPOINT = 'https://eth.api.onfinality.io/public';
+const HTTP_ENDPOINT = 'https://eth.llamarpc.com';
 const BLOCK_CONFIRMATIONS = 20;
 
 const ds: SubqlRuntimeDatasource = {
@@ -256,7 +257,7 @@ describe('Api.ethereum', () => {
 
     // Polygon
     ethApi = new EthereumApi(
-      'https://polygon.api.onfinality.io/public',
+      'https://polygon.llamarpc.com',
       BLOCK_CONFIRMATIONS,
       eventEmitter,
     );