+++ packages/node/src/ethereum/api.service.ethereum.test.ts
@@ -16,7 +16,8 @@ import { EthereumApiService } from './api.service.ethereum';
 
 // Add api key to work
 const WS_ENDPOINT = 'wss://eth.api.onfinality.io/ws?apikey=';
-const HTTP_ENDPOINT = 'https://eth.api.onfinality.io/public';
+// const HTTP_ENDPOINT = 'https://eth.api.onfinality.io/public';
+const HTTP_ENDPOINT = 'https://eth.llamarpc.com';
 
 function testSubqueryProject(endpoint: string): SubqueryProject {
   return {