+++ packages/node/src/subcommands/testing.module.ts
@@ -9,8 +9,10 @@ import {
   ConnectionPoolService,
   ConnectionPoolStateManager,
   DbModule,
+  InMemoryCacheService,
   NodeConfig,
   PoiService,
+  PoiSyncService,
   StoreCacheService,
   StoreService,
   TestRunner,
@@ -28,10 +30,12 @@ import { UnfinalizedBlocksService } from '../indexer/unfinalizedBlocks.service';
 
 @Module({
   providers: [
+    InMemoryCacheService,
     StoreService,
     StoreCacheService,
     EventEmitter2,
     PoiService,
+    PoiSyncService,
     SandboxService,
     DsProcessorService,
     DynamicDsService,