+++ packages/types/src/global.ts
@@ -2,6 +2,7 @@
 // SPDX-License-Identifier: GPL-3.0
 
 import {Provider} from '@ethersproject/abstract-provider';
+import '@subql/types-core/dist/global';
 
 declare global {
   const api: Provider;